1. Updating teacher_schedule timings_str depending upon how user fills the announced_timings for the program, when the program is getting announced. 2. If program is "full-day" then showing a teacher as blockable only if the teacher is available for the full-day.
Known Issues - TODO -
A. Fix bug in 1. above viz -- If a teacher is added to the program after it is announced, the teacher timings is not getting updated to program timings.
B. Fix bug in 2. above viz -- If a program is full-day, a teacher can still be scheduled for specific sessions across all days. Subsquently any attempt to add the same teacher for more sessions will fail since the teacher will not show in the blockable teacher list.